Refactor package path
