--- .idea/dictionaries/ripzery.xml
@@ -1,9 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="ripzery">
-    <words>
-      <w>idempotency</w>
-      <w>unfinalized</w>
-      <w>websocket</w>
-    </words>
-  </dictionary>
-</component>
