+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/CompositeSocketChannelListener.kt
@@ -8,7 +8,6 @@ package co.omisego.omisego.websocket.listener
  */
 
 import co.omisego.omisego.model.APIError
-import co.omisego.omisego.websocket.SocketChannelListener
 
 class CompositeSocketChannelListener(
     private val listeners: MutableSet<SocketChannelListener> = linkedSetOf()
