+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/CompositeSocketConnectionListener.kt
@@ -1,6 +1,11 @@
 package co.omisego.omisego.websocket.listener
 
-import co.omisego.omisego.websocket.SocketConnectionListener
+/**
+ * OmiseGO
+ *
+ * Created by Yannick Badoual on 7/13/2018 AD.
+ * Copyright © 2017-2018 OmiseGO. All rights reserved.
+ */
 
 /**
  * OmiseGO
