+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/listener/SocketConnectionListener.kt
@@ -1,4 +1,4 @@
-package co.omisego.omisego.websocket
+package co.omisego.omisego.websocket.listener
 
 /*
  * OmiseGO
