+++ omisego-sdk/src/test/kotlin/co/omisego/omisego/websocket/CompositeSocketConnectionListenerTest.kt
@@ -7,6 +7,8 @@ package co.omisego.omisego.websocket
  * Copyright © 2017-2018 OmiseGO. All rights reserved.
  */
 
+import co.omisego.omisego.websocket.listener.CompositeSocketConnectionListener
+import co.omisego.omisego.websocket.listener.SocketConnectionListener
 import com.nhaarman.mockito_kotlin.times
 import com.nhaarman.mockito_kotlin.verify
 import org.amshove.kluent.mock
