+++ omisego-sdk/src/test/kotlin/co/omisego/omisego/websocket/channel/dispatcher/SocketDispatcherTest.kt
@@ -10,8 +10,8 @@ package co.omisego.omisego.websocket.channel.dispatcher
 import co.omisego.omisego.custom.retrofit2.executor.MainThreadExecutor
 import co.omisego.omisego.model.socket.SocketReceive
 import co.omisego.omisego.utils.Either
-import co.omisego.omisego.websocket.SocketConnectionListener
-import co.omisego.omisego.websocket.SocketCustomEventListener
+import co.omisego.omisego.websocket.listener.SocketConnectionListener
+import co.omisego.omisego.websocket.listener.SocketCustomEventListener
 import co.omisego.omisego.websocket.enum.SocketCustomEvent
 import co.omisego.omisego.websocket.enum.SocketSystemEvent
 import com.nhaarman.mockito_kotlin.times
